JPA CRUD+Derived query params
jpa repository para crud y derived methos
se aplico DTO para el ingreso de datos(buena practica)
dto de salida para el Libro para que muetre info completa de libro + solo nombre del autor y no todo, para el autor no es necesario porque pues no hay info que se pueda ocultar o resumir
solo se maneja una relacion direccional libro tiene autor, Autor tiene varios libros no por mayor facilidad :)
aplique jsed para manejo de resultados (se puede mejorar los mensajitos y eso)

+++ src/main/java/com/example/biblioteca/service/AutorService.java
@@ -3,6 +3,7 @@ package com.example.biblioteca.service;
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.example.biblioteca.model.Autor;
 import com.example.biblioteca.repository.AutorRepository;
@@ -10,30 +11,54 @@ import com.example.biblioteca.repository.AutorRepository;
 @Service
 public class AutorService {
 
-    private final AutorRepository autorRepository;
+    @Autowired
+    private AutorRepository autorRepository;
 
-    //constructor
-    public AutorService(AutorRepository autorRepository) {
-        this.autorRepository = autorRepository;
-    }
-
-    //crear o actualizar
-    public Autor saveAutor(Autor autor) {
+    // crear
+    public Autor guardarAutor(Autor autor) {
         return autorRepository.save(autor);
     }
 
-    //listar
-    public List<Autor> getAllAutores() {
+    // obtener todos
+    public List<Autor> getAutores() {
         return autorRepository.findAll();
     }
 
-    //buscar por id
+    // buscar por id
     public Optional<Autor> getAutorById(Long id) {
         return autorRepository.findById(id);
     }
 
-    //eliminar por id
-    public void deleteAutor(Long id) {
-        autorRepository.deleteById(id);
+    // actualizar actuor por id
+    public Optional<Autor> actualizarAutor(Long id, Autor autorDetalles) {
+        // autor existe?
+        return autorRepository.findById(id)
+                // como devuelve un optional, si esta presente map se ejecuta sino devuelve
+                // opt.empty
+                .map(autorExistente -> {
+                    autorExistente.setNombre(autorDetalles.getNombre());
+                    autorExistente.setApellido(autorDetalles.getApellido());
+                    autorExistente.setNacionalidad(autorDetalles.getNacionalidad());
+                    return autorRepository.save(autorExistente);
+                });
+    }
+
+    // eliminar por id
+    public boolean eliminarPorId(Long id) {
+        if (autorRepository.existsById(id)) {
+            autorRepository.deleteById(id);
+            return true;
+        }
+        return false;
+    }
+
+    //DERIVED QUERIES METHODS
+
+    public List<Autor>buscarPorNacionalidad(String nacionalidad){
+        return autorRepository.findByNacionalidad(nacionalidad);
+    }
+
+    public List<Autor>buscarPorApellido(String parteDelApellido){
+        return autorRepository.findByApellidoContaining(parteDelApellido);
     }
 }
